Rename filebox API
