+++ cybsi/cloud/files/__init__.py
@@ -1,7 +1,6 @@
 """Use this section of API to with files.
 """
 
-from .api import FileboxAPI, FileboxAsyncAPI
 from .files import (
     FilesAPI,
     FilesAsyncAPI,
