+++ examples/file_upload_parts.py
@@ -2,7 +2,7 @@
 import os
 
 from cybsi.cloud import Client, Config
-from cybsi.cloud.filebox import LimitedReader
+from cybsi.cloud.files import LimitedReader
 
 if __name__ == "__main__":
     part_size = 10 * (1 << 20)  # 10mb
@@ -18,14 +18,14 @@ if __name__ == "__main__":
 
     config = Config(api_key="the cryptic string")
     with Client(config) as client:
-        session = client.filebox.files.create_session(part_size=part_size)
+        session = client.files.create_session(part_size=part_size)
         with open(file_path, "rb") as f:
             for part_number, part_size in enumerate(iter_part_size(), start=1):
-                client.filebox.files.upload_session_part(
+                client.files.upload_session_part(
                     LimitedReader(f, limit=part_size),
                     session_id=session.id,
                     part_number=part_number,
                     size=part_size,
                 )
-        ref = client.filebox.files.complete_session(session.id)
+        ref = client.files.complete_session(session.id)
         print(ref.id)
